Add Express backend with modular routes
