+++ package.json
@@ -18,6 +18,7 @@
     "open:android": "npx cap open android"
   },
   "dependencies": {
+    "@capacitor/core": "^5.0.6",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@hookform/resolvers": "^3.9.0",
@@ -55,6 +56,7 @@
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
     "embla-carousel-react": "^8.3.0",
+    "express": "^4.21.2",
     "i18next": "^23.10.1",
     "input-otp": "^1.2.4",
     "lucide-react": "^0.462.0",
@@ -74,10 +76,11 @@
     "tailwindcss-animate": "^1.0.7",
     "vaul": "^0.9.3",
     "zod": "^3.23.8",
-    "zustand": "^4.5.2",
-    "@capacitor/core": "^5.0.6"
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
+    "@capacitor/android": "^5.0.6",
+    "@capacitor/cli": "^5.0.6",
     "@eslint/js": "^9.9.0",
     "@tailwindcss/typography": "^0.5.15",
     "@testing-library/jest-dom": "^6.4.2",
@@ -100,8 +103,6 @@
     "typescript-eslint": "^8.0.1",
     "vite": "^5.4.1",
     "vite-plugin-pwa": "^1.0.0",
-    "vitest": "^1.5.0",
-    "@capacitor/cli": "^5.0.6",
-    "@capacitor/android": "^5.0.6"
+    "vitest": "^1.5.0"
   }
 }
