+++ server/routes/syncStatus.js
@@ -1,0 +1,9 @@
+import { Router } from "express";
+
+export default function createSyncStatusRouter({ getStatus }) {
+  const router = Router();
+  router.get("/", (req, res) => {
+    res.json(getStatus());
+  });
+  return router;
+}
